Cwebp - Part I
